✨ (blog): Add blog post about the T Coronae Borealis nova
This commit adds a new blog post about the T Coronae Borealis nova, which is expected to occur in September 2024. The blog post provides information about what a nova is, how to observe it, and where to find more information. It also includes a new image of a red giant star.